feat: add aidev ops scripts and make MySQL migrations deployable
Ship laptop→aidev provision/deploy/health for lms.pamungkas.org.
Trust Caddy proxies, shorten over-long MySQL identifiers, and type
last_lesson_id as unsigned bigint so the lessons FK can be created.

+++ database/migrations/2025_11_26_193402_create_enrollments_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->timestamp('started_at')->nullable();
             $table->timestamp('completed_at')->nullable();
             $table->foreignId('invited_by')->nullable()->constrained('users')->nullOnDelete();
-            $table->integer('last_lesson_id')->nullable();
+            $table->unsignedBigInteger('last_lesson_id')->nullable();
             $table->timestamps();
 
             $table->unique(['user_id', 'course_id']);
